Update CMake to include `entt.h`
Previously, CMake was configured to include only raylib.

`EnTT` provides a single include file (`entt.hpp`) that provides ECS
functionality.
This file was added to the project as a raw header file.
So, CMake needs to know where to find `entt.h`

Note:  prefer .h file extension. So, the extension was changed when
adding `EnTT` header to project.