Refactor variable names to free `game`
Previously, the name of the executable was simply set to `game`.

A `game` entity is being added to the source files.

Prefer that `game` only refers to the entity because that is less
ambiguous than `game` referring to the executable and the entity.

Substitute instances of `game` with `${PROJECT_NAME}`, which is
`boxtrap` in this case.